requirements.txt
openai>=1.0.0
pillow>=10.0.0
xai-sdk>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0
